+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # QOwnNotes Changelog
 
+## 0.66
+- fixed a problem with storing wrong todo list urls in the settings
+
 ## 0.65
 - you are now able to create new todo items on your ownCloud server
 - you are now able to delete todo items on your ownCloud server